Language: ? added, words changed.
Quite a hefty oneliner, so diff can't be shown in online editor. This is the change:

s/Are you sure you want to delete "{folderName}" and all files inside. This operation can not be undone/Are you sure you want to delete "{folderName}" and all files in it. This can not be undone/