Fix file drop shared folders
When ACL is enabled the READ permission is required to get any other
permission on the group folder (or its contents). However, the
ACLCacheWrapper verified not only that the READ permission was enabled
for the folder/file, but also that it was being requested. Due to this
it was not possible to publicly upload a file to a file drop shared
folder inside a group folder, as in that case only SHARE + CREATE
permissions are requested.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>