Fix wrong object passed to the mount provider in VersionsBackend::getAllVersionedFiles
`$folder['groups']` is an array from groupId to permission. The mount
provider only wants the permission, so since we want to get all the
versionned files, just use 31 instead.

Before php8, this code was working since using & for an int and an array
gave an int. In php8, this now thrown an error.

This also include more type hinting that helped me figure out that was
wrong.

Fix #1725:

Signed-off-by: Carl Schwan <carl@carlschwan.eu>