feat(React, Controller, Service, Route): Display or not the Groupfolder admin delegation
Now, we can display the Groupfolder admin delegation with 2 conditions : 1) If the current user is nextcloud admin AND 2) If the 'workspace' app is installed.

To check if an app is installed I created the const APPS_USE_GROUPFOLDERS where a dev can fill this list to check if all apps are installed.

Signed-off-by: Baptiste Fotia <fotia.baptiste@hotmail.com>