refactor(): Change the API/REST to update groups
When we update delegation groups, we use this route :
/settings/authorizedgroups/saveSettings .
One more, I apply the DRY principle and now, it's the same function to update and
get the list of delegations groups.
We need to add the class name as a parameter to specify whether it's for the admin
or the subadmin.

Please, read the Carl's comment : https://github.com/nextcloud/groupfolders/pull/2072#discussion_r981404761

Signed-off-by: Baptiste Fotia <fotia.baptiste@hotmail.com>