tests: Add E2E tests for copying and moving shared groupfolders
New filesUtils functions adjusted from
cypress/e2e/files_sharing/filesSharingUtils.ts of Nextcloud server at
commit 30f3cbf5f8578ef8cb6a77ecd304df83089985f0; note that
"shareSettings" is not used, so there is no need to include the hack in
"createShare" nor add the "updateShare" function.

Besides that the helper functions to copy and move a file now support
using as destination a directory that is not a descendant of the current
directory (as that requires to first navigate to the Home folder in the
file picker, which was not supported before, to then descend to the
final destination).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>