Fix slow move on same object bucket during file upload.
This commit fixes the issue https://github.com/nextcloud/server/issues/47856. When you upload a file into a group folder and when you use a single S3 bucket as primary storage, the final move operation hangs for a long time. In the background, Nextcloud initiates a copy-delete sequence from the bucket into the bucket, with causes a lot unnecessary overhead. Nextcloud thinks that the file must be imported to another storage and does not recognize that everything is done on the same object bucket. In that case, the import step can be completely skipped, which saves time, network bandwidth and reduces the load on the object storage.

Signed-off-by: Christoph Fiehe <c.fiehe@eurodata.de>